Eahw 1764/refactor (#12)
* Delegate whole responsibility of accrual calculation to Accrual Modules, passing on whole time entry, instead of daily tranches
Only patch accruals if list isn't empty

* Handle case when accrual placeholders for a given type aren't in place

* Delete empty file

---------
